refactor: root tab
