+++ src/app/[locale]/(viewer)/viewer/@news/_news-stack/client.tsx
@@ -1,6 +1,7 @@
 // import Image from "next/image";
 // import defaultOgImage from "@/assets/defaultOgImage.jpg";
 
+import { LinkCard } from "@/components/card/link-card";
 import {
 	Card,
 	CardContent,
@@ -26,20 +27,30 @@ export const NewsStackClient = ({ data }: Props) => {
 			<div className="m-2 space-y-4">
 				{data.map((d) => {
 					return (
-						<Card id={d.url} key={d.url}>
-							<CardHeader>
-								<CardTitle className="text-primary-grad">{d.title}</CardTitle>
-								<CardDescription>{d.quote}</CardDescription>
-							</CardHeader>
-							<CardContent>
-								<div>
-									{/* TODO: change to use fragment */}
-									{/* <Fragment set:html={sanitizeHtml(d.quote ?? "")} /> */}
-								</div>
-								<a href={d.url} rel="noopener noreferrer" target="_blank">
-									<div className="flex w-full justify-start">
-										{/* TODO: show images */}
-										{/* {d.ogImageUrl === null ? (
+						<>
+							<LinkCard
+								data={{
+									id: d.id,
+									badgeText: d.title,
+									title: d.title,
+									href: d.url,
+								}}
+								showDeleteButton={false}
+							/>
+							<Card id={d.url} key={d.url}>
+								<CardHeader>
+									<CardTitle className="text-primary-grad">{d.title}</CardTitle>
+									<CardDescription>{d.quote}</CardDescription>
+								</CardHeader>
+								<CardContent>
+									<div>
+										{/* TODO: change to use fragment */}
+										{/* <Fragment set:html={sanitizeHtml(d.quote ?? "")} /> */}
+									</div>
+									<a href={d.url} rel="noopener noreferrer" target="_blank">
+										<div className="flex w-full justify-start">
+											{/* TODO: show images */}
+											{/* {d.ogImageUrl === null ? (
 									<Image
 										alt=""
 										className="hidden size-32 object-cover sm:block"
@@ -55,16 +66,19 @@ export const NewsStackClient = ({ data }: Props) => {
 										width="128"
 									/>
 								)} */}
-										<div className="w-full p-4">
-											<CardTitle className="break-words">{d.ogTitle}</CardTitle>
-											<CardDescription className="break-words">
-												{d.ogDescription}
-											</CardDescription>
+											<div className="w-full p-4">
+												<CardTitle className="break-words">
+													{d.ogTitle}
+												</CardTitle>
+												<CardDescription className="break-words">
+													{d.ogDescription}
+												</CardDescription>
+											</div>
 										</div>
-									</div>
-								</a>
-							</CardContent>
-						</Card>
+									</a>
+								</CardContent>
+							</Card>
+						</>
 					);
 				})}
 			</div>
