+++ src/components/root-tab.stories.tsx
@@ -1,10 +1,10 @@
 import type { Meta, StoryObj } from "@storybook/nextjs-vite";
-import { RootTab } from "./client";
+import { RootTab } from "./root-tab";
 
 const meta = {
 	component: RootTab,
-	tags: ["autodocs"],
 	parameters: { layout: "fullscreen" },
+	tags: ["autodocs"],
 } satisfies Meta<typeof RootTab>;
 
 export default meta;
@@ -14,8 +14,8 @@ type Story = StoryObj<typeof meta>;
 export const Default: Story = {
 	args: {
 		news: <div>sample news</div>,
-		books: <div>sample books</div>,
 		contents: <div>sample contents</div>,
-		images: <div>sample image</div>,
+		images: <div>sample images</div>,
+		books: <div>sample books</div>,
 	},
 };
